Add __initialize_board() method to Gameboard
Added the __initialize_board() method to the Gameboard class to create
the initial board. This change has been made because the board now
stores Square objects instead of strings for the pieces. As such,
initializing the board is more complex and deserves a separate function.

As a consequence, also deleted the print_board() function since it has
to be rewritten.