Add BoardInfo class
Added a BoardInfo class to store board-related information such as the
file letters and rank numbers.

This addition was made to avoid class coupling. As it stands,
is_valid_square() is a method of MoveValidator but this will change
soon when it is moved to Square. Since Square is used in Gameboard, and
the implementation of is_valid_square() uses Gameboard, this will lead
coupling which is undesirable. The introduction of BoardInfo aims to
solve this problem.