Make all imports of user-defined modules relative
When a tests/ directory was added containing a test_square.py file to
unit test the is_valid_square() method of the Square class, PyTest
raised a ModuleNotFoundError saying that the board_info module imported
by square.py in cli_chess/ could not be found. Whatever the explanation
for this error is, it turns out that it could be fixed by making the
import relative.

Presumably, the explanation is that whenever Python imports a file that
imports another file, it thinks the second import is relative to the
original file, not the imported file. In any case, to prevent this error
from occuring again all imports of user-defined modules were made
relative.