Add play_move() method to Gameboard
Add the play_move() method to the Gameboard class to move a piece from
one square to another. Note that the function expects correct input (i.e
the move is a valid move, the move can be played legally, etc.). The
responsibility for validating the input is left to the class that will
control the game flow.